Make RidgeClassifier output more like probabilities
As described in Returns the distance from the hyperplane of the
RidgeClassifer. Not a probability, per se, but values are still rankable for
LPOCV.

        Probability for a pair looks like this:
        array([[0.56079167, 0.43920833],
               [0.562586  , 0.437414  ]])

The decision function produces distance from hyperplane for each
sample. predict_proba usually produces one set of probabilities per
sample
[p_class_A, p_class_B] where p_class_A = 1-p_class_B

To produce something similar to this, negate the distances.
Append negated values, and return transposed array.
[dist_hyperplane, -dist_hyperplane]